move smt options where they are accessible by kompile
